Move the function to time passes to util
